+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [Bug #31] Detect secrets assigned with := operator
 - Fix pre-commit getting installed at wrong filename
 - Added more patterns to search for secret assignments
+- Added --only-matching option
 
 ## 0.1.3 (2022-04-22)